Added script prove-with-testserver and example webtest
The example selenium test (90-selenium-webtests.t) will be reduced in the
future by moving most of the code to a custom module, but this shows the
boilerplate needed for a webtest, for now.

I also added a script "script/prove-with-testserver.pl" which starts an
instance of the webapp, runs the tests with "prove" (passing through any
arguments) and then shuts it down at the end of the tests.