Wrote a basic test of the login/logout sequence.
This testcase is a good example of how to browse pages and find elements
and interact with them.  It performs the login by clicking the login
link, clicking the username field, and then using keyboard shortcuts
(tab and enter) to put in a wrong password, and submit the form, then
try again with a correct password, then log out using the logout link.

Also added a feature to Selenium::TestUtil that automatically closes
the browser session at the end of the script, unless SELENIUM_NOCLOSE
is set.